periodic update TestCaseDesign
