+++ reservation-service/src/test/java/com/revature/service/ReservationServiceApplicationTests.java
@@ -7,31 +7,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 class ReservationServiceApplicationTests {
 
 	
-	Reservation testRes;
-	
 	@Test
-    void isValidReservation(Reservation r, Location rmId) {
-		//check to see if room is available for a  
-		//reservation during proposed time slot
-
-	
-	}
-	
-	@Test
-	void isInvalidReservation(Reservation r, Location rmId) {
+	public void contextLoads() {
 		
 	}
-	@Test
-	void makeReservation(Reservation r, Location rmId) {
-		//see if making a reservation was successful
-	}
-	
-	@Test
-	void cancelReservation(Reservation r, Location rmId) {
-		//check to see if removal of reservation was successful
-	}
-	
-	
 	
 
 }
