+++ reservation-service/src/test/java/com/revature/service/TestMakeReservationTimeSlot.java
@@ -1,0 +1,40 @@
+package com.revature.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class TestMakeReservationTimeSlot {
+
+	
+Reservation testRes;
+	
+	@Test
+    void isValidReservation(Reservation r, Location rmId) {
+		//check to see if room is available for a  
+		//reservation during proposed time slot
+
+	
+	}
+	
+	@Test
+	void isInvalidReservation(Reservation r, Location rmId) {
+		
+	}
+	@Test
+	void makeReservation(Reservation r, Location rmId) {
+		//see if making a reservation was successful
+	}
+	
+	@Test
+	void cancelReservation(Reservation r, Location rmId) {
+		//check to see if removal of reservation was successful
+	}
+	
+	
+	@Test
+	void test() {
+		fail("Not yet implemented");
+	}
+
+}
